refactor: 모두 체크 표시를 위한 스토어 구독
