+++ frontend/src/components/NewChecklist/ChecklistTab/NewChecklistTab.tsx
@@ -2,19 +2,21 @@ import { useMemo } from 'react';
 
 import ChecklistTabFallback from '@/components/_common/errorBoundary/ChecklistTabFallback';
 import Tabs from '@/components/_common/Tabs/Tabs';
-import useGetChecklistQuestionQuery from '@/hooks/query/useGetChecklistQuestionQuery';
+import useInitialChecklist from '@/hooks/useInitialChecklist';
 import useTabs from '@/hooks/useTabs';
+import useChecklistStore from '@/store/useChecklistStore';
 
 const NewChecklistTab = () => {
-  const { data: checklist, isFetched, isLoading } = useGetChecklistQuestionQuery();
+  const { data: checklist, isFetched, isLoading } = useInitialChecklist();
+  const checklistStore = useChecklistStore(state => state.checklistCategoryQnA);
   const { getTabsForChecklist } = useTabs();
 
   const categoryTabs = useMemo(() => {
-    if (isFetched && checklist) {
+    if (isFetched && checklist && checklistStore.length) {
       return getTabsForChecklist(checklist);
     }
     return [];
-  }, [isFetched, getTabsForChecklist]);
+  }, [isFetched, getTabsForChecklist, checklistStore]);
 
   if (isLoading) return <ChecklistTabFallback />;
